Delete Tiktok.txt
This was obsolete. There are a lot more domains than what I was blocking and I realized I already used other blocklists which had waaay more. This is supposed to be a collection of additional sites which may not already be patched on my network, or yours, which complement my threat model and patch whatever other blocklists don't. Use other lists if you want TikTok blocked.